requirements.txt
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
requests==2.32.3
Pillow==10.4.0
python-decouple==3.8
dj-database-url==2.2.0
whitenoise==6.6.0 